requirements/problems.txt
djangospam==1.0.0
django-registration==1.0
django-picklefield==0.3.0
django-axes==1.3.3 ???